requirements.txt
Flask==1.0.2
simplejson==3.16.0
python-dotenv==0.9.1
requests==2.20.0
gunicorn==19.9.0
gevent==1.3.7
biopython==1.75
